Add files via upload
This is an ongoing project of mine. It is a Time Card Application meant to be used as a Clock In/Clock Out program to help employees keep track of hours employees are working without the use of physical time cards! It is connected to an SQL Database Server where it stores all of the employees names, when they clock in/clock out, and the Date! I will be including any SQL files that I can for examples. I am including a text file exported from Microsoft Sql Server Management Studio of a Table I created inside of a database on my SQL server. As well as an example of SQL query I have been using to navigate my table. More files will be added as I finish this project.