Add files via upload
This is an ongoing personal project. It is a C# Time Card Forms Application! It's main purpose, is to keep track of Clock In and Clock Out times of employees at any workplace! I used an SQL server and SQL database to store all the employees information, clock in and out times, etc. I have included an example of the table I created, that was exported into a text file from the SQL database. As well as an example of the Query commands I used to navigate my table and database contents. Going forward I would like to add a second form that is for Employer use that can help them with payroll information. More files will be added as I finish the project.