Add files via upload
This is my Python Movie Library Database Project!!! This project took so much research. I was not expecting to have to learn as much as I did to complete this. It is a simple idea using tkinter library to create a Python Form Application. To summarize, the project connects to a SQL data base and table, and allows the user to look up any movie within that table! The user has the choice of searching by movie name, movie rating, genre, and year released! While 10 movies isn't a lot to work with. I do believe this application would work with larger databases! I wish I could have imported 1000 movies and their information efficiently. I learned so much about python and tkinter through this project. I am very proud of it! 

While the project isn't visually stunning. I do believe this is the best python code I have written to date! I hope it is as Readable as I wanted it to be! I provided a SQL file of queries I used throughout the project. As well as a table I exported from Microsoft SQL Server Management Studio.